Refactor chat store to separate file
